+++ src/macros_from_to.rs
@@ -1236,20 +1236,20 @@ type Fix = fixed::",
             $s_fixed,
             "<fixed::frac::U4>;
 // 1.75 is 1.11 in binary
-let f = Fix::from_str_bin(\"1.11\");
+let f = Fix::from_str_binary(\"1.11\");
 let check = Fix::from_bits(0b111 << (4 - 2));
 assert_eq!(f, Ok(check));
 ",
             if_signed_else_empty_str!(
                 $Signedness,
-                "let neg = Fix::from_str_bin(\"-1.11\");
+                "let neg = Fix::from_str_binary(\"-1.11\");
 assert_eq!(neg, Ok(-check));
 ",
             ),
             "```
 ";
             #[inline]
-            pub fn from_str_bin(src: &str) -> Result<$Fixed<Frac>, ParseFixedError> {
+            pub fn from_str_binary(src: &str) -> Result<$Fixed<Frac>, ParseFixedError> {
                 FromStrRadix::from_str_radix(src, 2)
             }
         );